fix(diff.py): 尝试消除循环引用
